records/requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
babel==2.17.0
beautifulsoup4==4.13.3
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
cobble==0.1.4
colorama==0.4.6
cryptography==43.0.3
cssselect==1.3.0
cssselect2==0.8.0
Django==5.1.6
django-allauth==65.4.1
django-apscheduler==0.7.0
django-crispy-forms==2.3
django-environ==0.12.0
django-suit==0.2.28
djangorestframework==3.15.1
docx2pdf==0.1.8
docx2python==3.5.0
docxcompose==1.4.0
docxtpl==0.19.1
fonttools==4.56.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gunicorn
httplib2==0.22.0
idna==3.7
Jinja2==3.1.6
lxml==5.3.0
mammoth==1.9.0
MarkupSafe==3.0.2
oauthlib==3.2.2
paragraphs==1.0.1
pdfrw==0.4
pillow==11.0.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.21.0
pydyf==0.11.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyPDF2==3.0.1
pyphen==0.17.2
python-docx==1.1.2
python-dotenv==1.0.1
qrcode==8.0
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==76.1.0
six==1.17.0
soupsieve==2.5
sqlparse==0.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
types-html5lib==1.1.11.20241018
types-lxml==2025.3.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.2.1
weasyprint==64.1
webencodings==0.5.1
zopfli==0.2.3.post1